requirements.txt
Django==4.0.9
djangorestframework==3.14.0
gspread==5.9.0
oauth2client==4.1.3
gevent==22.10.2
gunicorn==20.1.0
